fix(types): 修复 CI mypy strict + ruff format 失败
mypy (13 errors → 0):
- portfolio/optimizer: register_optimizer 返回类型改为 Callable 装饰器签名
  (原标注 type[WeightOptimizer] 导致 4 个子类 Too many arguments)
- factor/engine: _datetime_to_int 用 isinstance 收窄替代 object→int 强转
- factor/analysis: 删多余 type:ignore（改由 mypy override 统一处理 scipy）
- backtest/orders, execution: np.sqrt 表达式用 float() 包裹消除 no-any-return
- MyTT.pyi: MACD 签名删除错误的 LOW/HIGH 参数（与 MyTT.py 实际签名对齐）
- pyproject: 新增 scipy mypy override (ignore_missing_imports)

ruff format: 8 个 test 文件格式化

验证: 564 passed, mypy 192 文件零错误, ruff check/format 全绿

+++ src/easy_tdx/backtest/orders.py
@@ -326,7 +326,7 @@ class OrderSimulator:
         if len(returns) < 2:
             return 0.0
         daily_vol = float(np.std(returns))
-        return daily_vol * np.sqrt(252)
+        return float(daily_vol * np.sqrt(252))
 
     def _execute_buy(
         self,
